requirements.txt
streamlit
pymupdf
langchain
langchain-google-genai
google-generativeai
fpdf2
